+++ cmd/template/command.go
@@ -7,12 +7,14 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
+	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/giantswarm/kubectl-gs/cmd/template/app"
 	"github.com/giantswarm/kubectl-gs/cmd/template/catalog"
 	"github.com/giantswarm/kubectl-gs/cmd/template/cluster"
 	"github.com/giantswarm/kubectl-gs/cmd/template/networkpool"
 	"github.com/giantswarm/kubectl-gs/cmd/template/nodepool"
+	"github.com/giantswarm/kubectl-gs/cmd/template/organization"
 )
 
 const (
@@ -22,6 +24,9 @@ const (
 
 type Config struct {
 	Logger micrologger.Logger
+
+	K8sConfigAccess clientcmd.ConfigAccess
+
 	Stderr io.Writer
 	Stdout io.Writer
 }
@@ -71,6 +76,9 @@ func New(config Config) (*cobra.Command, error) {
 	{
 		c := cluster.Config{
 			Logger: config.Logger,
+
+			K8sConfigAccess: config.K8sConfigAccess,
+
 			Stderr: config.Stderr,
 			Stdout: config.Stdout,
 		}
@@ -109,6 +117,19 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
+	var organizationCmd *cobra.Command
+	{
+		c := organization.Config{
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		organizationCmd, err = organization.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	f := &flag{}
 
 	r := &runner{
@@ -132,6 +153,7 @@ func New(config Config) (*cobra.Command, error) {
 	c.AddCommand(clusterCmd)
 	c.AddCommand(networkpoolCmd)
 	c.AddCommand(nodepoolCmd)
+	c.AddCommand(organizationCmd)
 
 	return c, nil
 }