+++ cmd/template/nodepool/provider/azure.go
@@ -21,10 +21,69 @@ import (
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	"sigs.k8s.io/yaml"
 
+	azurenodepooltemplate "github.com/giantswarm/kubectl-gs/cmd/template/nodepool/provider/templates/azure"
 	"github.com/giantswarm/kubectl-gs/internal/key"
 )
 
 func WriteAzureTemplate(out io.Writer, config NodePoolCRsConfig) error {
+	var err error
+
+	if key.IsCAPZVersion(config.ReleaseVersion) {
+		err = WriteCAPZTemplate(out, config)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	} else {
+		err = WriteGSAzureTemplate(out, config)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	return nil
+}
+
+func WriteCAPZTemplate(out io.Writer, config NodePoolCRsConfig) error {
+	var err error
+
+	data := struct {
+		KubernetesVersion  string
+		ClusterName        string
+		Description        string
+		MaxSize            int
+		MinSize            int
+		Name               string
+		Namespace          string
+		Organization       string
+		Replicas           int
+		StorageAccountType string
+		Version            string
+		VMSize             string
+	}{
+		KubernetesVersion:  "v1.19.9",
+		ClusterName:        config.ClusterName,
+		Description:        config.Description,
+		MaxSize:            config.NodesMax,
+		MinSize:            config.NodesMin,
+		Name:               config.NodePoolID,
+		Namespace:          key.OrganizationNamespaceFromName(config.Organization),
+		Organization:       config.Organization,
+		Replicas:           config.NodesMin,
+		StorageAccountType: key.AzureStorageAccountTypeForVMSize(config.VMSize),
+		Version:            config.ReleaseVersion,
+		VMSize:             config.VMSize,
+	}
+
+	t := template.Must(template.New(config.FileName).Parse(azurenodepooltemplate.GetTemplate()))
+	err = t.Execute(out, data)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+func WriteGSAzureTemplate(out io.Writer, config NodePoolCRsConfig) error {
 	var err error
 
 	var azureMachinePoolCRYaml, machinePoolCRYaml, sparkCRYaml []byte
@@ -101,7 +160,7 @@ func newAzureMachinePoolCR(config NodePoolCRsConfig) *expcapzv1alpha3.AzureMachi
 				label.Cluster:                 config.ClusterName,
 				capiv1alpha3.ClusterLabelName: config.ClusterName,
 				label.MachinePool:             config.NodePoolID,
-				label.Organization:            config.Owner,
+				label.Organization:            config.Organization,
 			},
 		},
 		Spec: expcapzv1alpha3.AzureMachinePoolSpec{