+++ internal/key/key.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	semver "github.com/blang/semver/v4"
+	"github.com/blang/semver/v4"
 	"github.com/giantswarm/microerror"
 	"github.com/spf13/afero"
 	v1 "k8s.io/api/core/v1"
@@ -129,6 +129,11 @@ func IsCAPAVersion(version string) bool {
 	return version == "20.0.0"
 }
 
+// IsCAPZVersion returns whether a given GS Release Version is based on the CAPI/CAPZ projects
+func IsCAPZVersion(version string) bool {
+	return version == "20.0.0"
+}
+
 // IsOrgNamespaceVersion returns whether a given AWS GS Release Version is based on clusters in Org Namespace
 func IsOrgNamespaceVersion(version string) bool {
 	// TODO: this has to return true as soon as v16.0.0 is the newest version
@@ -136,7 +141,7 @@ func IsOrgNamespaceVersion(version string) bool {
 	// see https://github.com/giantswarm/aws-admission-controller/blob/ef83d90fc856fbc0484bec967064834c0b8d2c1e/pkg/aws/v1alpha3/cluster/mutate_cluster.go#L191-L202
 	// so as soon as the latest version is >=16.0.0 we are going to need the org-namespace as default here.
 	if version == "" {
-		return false
+		return true
 	}
 	OrgNamespaceVersion, _ := semver.New(FirstAWSOrgNamespaceRelease)
 	releaseVersion, _ := semver.New(version)
@@ -214,4 +219,31 @@ func OrganizationNamespaceFromName(name string) string {
 	name = normalize.AsDNSLabelName(fmt.Sprintf(organizationNamespaceFormat, name))
 
 	return name
+}
+
+func AzureStorageAccountTypeForVMSize(vmSize string) string {
+	// Ugly but hardcoding is the only way to know premium storage support without reaching Azure API.
+	vmTypesWithPremiumStorageSupport := map[string]bool{
+		"Standard_B12ms": true, "Standard_B16ms": true, "Standard_B20ms": true,
+		"Standard_F2s": true, "Standard_F4s": true, "Standard_F8s": true, "Standard_F16s": true,
+		"Standard_D4s_v3": true, "Standard_D8s_v3": true, "Standard_D16s_v3": true, "Standard_D32s_v3": true, "Standard_D48s_v3": true, "Standard_D64s_v3": true,
+		"Standard_E4-2s_v3": true, "Standard_E4s_v3": true, "Standard_E8-2s_v3": true, "Standard_E8-4s_v3": true, "Standard_E8s_v3": true, "Standard_E16-4s_v3": true, "Standard_E16-8s_v3": true,
+		"Standard_E16s_v3": true, "Standard_E20s_v3": true, "Standard_E32-8s_v3": true, "Standard_E32-16s_v3": true, "Standard_E32s_v3": true,
+		"Standard_PB6s":     true,
+		"Standard_E4-2s_v4": true, "Standard_E4s_v4": true, "Standard_E8-2s_v4": true, "Standard_E8-4s_v4": true, "Standard_E8s_v4": true, "Standard_E16-4s_v4": true, "Standard_E16-8s_v4": true, "Standard_E16s_v4": true, "Standard_E20s_v4": true, "Standard_E32-8s_v4": true, "Standard_E32-16s_v4": true, "Standard_E32s_v4": true, "Standard_E48s_v4": true, "Standard_E64-16s_v4": true, "Standard_E64-32s_v4": true, "Standard_E64s_v4": true,
+		"Standard_E4-2ds_v4": true, "Standard_E4ds_v4": true, "Standard_E8-2ds_v4": true, "Standard_E8-4ds_v4": true, "Standard_E8ds_v4": true, "Standard_E16-4ds_v4": true, "Standard_E16-8ds_v4": true, "Standard_E16ds_v4": true, "Standard_E20ds_v4": true, "Standard_E32-8ds_v4": true, "Standard_E32-16ds_v4": true, "Standard_E32ds_v4": true, "Standard_E48ds_v4": true, "Standard_E64-16ds_v4": true, "Standard_E64-32ds_v4": true, "Standard_E64ds_v4": true,
+		"Standard_D4ds_v4": true, "Standard_D8ds_v4": true, "Standard_D16ds_v4": true, "Standard_D32ds_v4": true, "Standard_D48ds_v4": true, "Standard_D64ds_v4": true,
+		"Standard_D4s_v4": true, "Standard_D8s_v4": true, "Standard_D16s_v4": true, "Standard_D32s_v4": true, "Standard_D48s_v4": true, "Standard_D64s_v4": true,
+		"Standard_F4s_v2": true, "Standard_F8s_v2": true, "Standard_F16s_v2": true, "Standard_F32s_v2": true, "Standard_F48s_v2": true, "Standard_F64s_v2": true, "Standard_F72s_v2": true,
+		"Standard_NV12s_v3": true, "Standard_NV24s_v3": true, "Standard_NV48s_v3": true,
+		"Standard_L8s_v2": true, "Standard_L16s_v2": true, "Standard_L32s_v2": true, "Standard_L48s_v2": true, "Standard_L64s_v2": true, "Standard_L80s_v2": true,
+		"Standard_M208s_v2": true, "Standard_M416-208s_v2": true, "Standard_M416s_v2": true, "Standard_M416-208ms_v2": true, "Standard_M416ms_v2": true,
+		"Standard_NV4as_v4": true, "Standard_NV8as_v4": true, "Standard_NV16as_v4": true, "Standard_NV32as_v4": true,
+		"Standard_D4as_v4": true, "Standard_D8as_v4": true, "Standard_D16as_v4": true, "Standard_D32as_v4": true, "Standard_D48as_v4": true, "Standard_D64as_v4": true, "Standard_D96as_v4": true,
+		"Standard_E4as_v4": true, "Standard_E8as_v4": true, "Standard_E16as_v4": true, "Standard_E20as_v4": true, "Standard_E32as_v4": true, "Standard_E48as_v4": true, "Standard_E64as_v4": true, "Standard_E96as_v4": true,
+	}
+	if vmTypesWithPremiumStorageSupport[vmSize] {
+		return "Premium_LRS"
+	}
+	return "Standard_LRS"
 }