+++ pkg/installation/url_test.go
@@ -44,9 +44,9 @@ func Test_getBasePath(t *testing.T) {
 			errorMatcher: IsUnknownUrlType,
 		},
 		{
-			name:         "case 5: using giant swarm api url",
-			url:          "https://api.g8s.test.eu-west-1.aws.coolio.com",
-			errorMatcher: IsUnknownUrlType,
+			name:           "case 5: using giant swarm api url",
+			url:            "https://api.g8s.test.eu-west-1.aws.coolio.com",
+			expectedResult: "g8s.test.eu-west-1.aws.coolio.com",
 		},
 	}